[feat] API 명세 기반 검사 도메인 백엔드 정합화

+++ backend-spring/src/main/java/com/example/factoryguard/adapter/in/web/inspection/dto/InspectionRunResponse.java
@@ -3,6 +3,7 @@ package com.example.factoryguard.adapter.in.web.inspection.dto;
 import com.example.factoryguard.domain.inspection.model.RunStatus;
 import com.example.factoryguard.domain.inspection.model.RunType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record InspectionRunResponse(
@@ -12,7 +13,7 @@ public record InspectionRunResponse(
         Long targetId,
         RunType runType,
         RunStatus runStatus,
-        double appliedThreshold,
+        BigDecimal appliedThreshold,
         String errorCode,
         LocalDateTime startedAt,
         LocalDateTime completedAt
